archive/legacy-scripts/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
qdrant-client>=1.11.3
mistralai>=1.2.4
elevenlabs>=1.9.0
pydantic>=2.9.0
requests>=2.31.0
predicthq>=1.0.0
streamlit>=1.28.0

# MCP + Claude (hotel context layer)
anthropic>=0.40.0
mcp>=1.0.0

# Apaleo PMS integration
httpx>=0.27.0

# Voice input (Whisper STT) — optional, for demo voice mode
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.0